requirements.txt
ConfigArgParse==1.7
matplotlib==3.7.5
numpy==1.24.4
open_clip_torch==2.24.0
opencv_python==4.9.0.80
Pillow==9.5.0
scikit_learn==1.3.2
scipy==1.12.0
seaborn==0.13.2
timm==0.9.16
torch==1.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.66.2
utils==1.0.2
